Another improvements
+Improved destruction system
+Added no fuel image to busted screen
+Increased music level, decreased sfx level

+++ Assets/Scripts/DestroyMechanic/Debree.cs
@@ -33,7 +33,7 @@ public class Debree : MonoBehaviour
     {
         _rB.AddForce(direction, ForceMode.VelocityChange);
     }
-    public void AddExplosionForce(Vector3 origin,int destructionForceVal)
+    public void AddExplosionForce(Vector3 origin,float destructionForceVal)
     {
         AllPhysicalDebries.AddLast(this);
         AlreadyExploded = true;
